Added design for insert/update folder
